# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WalrusVault API - Decentralized file storage with SUI wallet auth
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FileRecord {
  blobId: string;
  fileName: string;
  size: number;
  /** @nullable */
  sha256Hash: string | null;
  uploadedAt: string;
  source: string;
  /** @nullable */
  walrusBlobId: string | null;
}

export interface FileUploadInput {
  fileName: string;
  fileData: string;
  walletAddress: string;
  signature: string;
}

export interface FileUploadResult {
  success: boolean;
  blobId: string;
  fileName: string;
  size: number;
  sha256Hash: string;
  message: string;
  /** @nullable */
  walrusBlobId: string | null;
  source: string;
}

export interface FileListResult {
  success: boolean;
  files: FileRecord[];
  count: number;
}

export interface FileDownloadInput {
  blobId: string;
  walletAddress: string;
  signature: string;
}

export interface FileDownloadResult {
  success: boolean;
  fileName: string;
  fileData: string;
  size: number;
  sha256Hash: string;
  verified: boolean;
  source: string;
  /** @nullable */
  walrusBlobId: string | null;
}

export interface FileDeleteInput {
  signature: string;
}

export interface DeleteResult {
  success: boolean;
  message: string;
}

export interface StorageStats {
  totalFiles: number;
  totalSize: number;
  totalSizeMb: number;
  /** @nullable */
  oldestUpload: string | null;
  /** @nullable */
  newestUpload: string | null;
  walrusFiles: number;
  localFiles: number;
}

export interface ActivityLogEntry {
  id: number;
  action: string;
  fileName: string;
  blobId: string;
  /** @nullable */
  size: number | null;
  createdAt: string;
  /** @nullable */
  source: string | null;
}

export interface ActivityLogResult {
  success: boolean;
  logs: ActivityLogEntry[];
  count: number;
}

export interface DriveImportInput {
  fileId: string;
  fileName: string;
  mimeType: string;
  accessToken: string;
  walletAddress: string;
  signature: string;
}