# lib/api-zod/src/generated/types/activityLogEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WalrusVault API - Decentralized file storage with SUI wallet auth
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLogEntry {
  id: number;
  action: string;
  fileName: string;
  blobId: string;
  /** @nullable */
  size: number | null;
  createdAt: string;
  /** @nullable */
  source: string | null;
}