# lib/api-zod/src/generated/types/activityLogResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WalrusVault API - Decentralized file storage with SUI wallet auth
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityLogEntry } from './activityLogEntry';

export interface ActivityLogResult {
  success: boolean;
  logs: ActivityLogEntry[];
  count: number;
}