# lib/api-zod/src/generated/types/fileDownloadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WalrusVault API - Decentralized file storage with SUI wallet auth
 * OpenAPI spec version: 0.1.0
 */

export interface FileDownloadInput {
  blobId: string;
  walletAddress: string;
  signature: string;
}