# lib/api-zod/src/generated/types/fileDownloadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WalrusVault API - Decentralized file storage with SUI wallet auth
 * OpenAPI spec version: 0.1.0
 */

export interface FileDownloadResult {
  success: boolean;
  fileName: string;
  fileData: string;
  size: number;
  sha256Hash: string;
  verified: boolean;
  source: string;
  /** @nullable */
  walrusBlobId: string | null;
}