# lib/api-zod/src/generated/types/fileListResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WalrusVault API - Decentralized file storage with SUI wallet auth
 * OpenAPI spec version: 0.1.0
 */
import type { FileRecord } from './fileRecord';

export interface FileListResult {
  success: boolean;
  files: FileRecord[];
  count: number;
}