# lib/api-zod/src/generated/types/fileRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WalrusVault API - Decentralized file storage with SUI wallet auth
 * OpenAPI spec version: 0.1.0
 */

export interface FileRecord {
  blobId: string;
  fileName: string;
  size: number;
  /** @nullable */
  sha256Hash: string | null;
  uploadedAt: string;
  source: string;
  /** @nullable */
  walrusBlobId: string | null;
}