# lib/api-zod/src/generated/types/fileUploadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WalrusVault API - Decentralized file storage with SUI wallet auth
 * OpenAPI spec version: 0.1.0
 */

export interface FileUploadResult {
  success: boolean;
  blobId: string;
  fileName: string;
  size: number;
  sha256Hash: string;
  message: string;
  /** @nullable */
  walrusBlobId: string | null;
  source: string;
}