# lib/api-zod/src/generated/types/storageStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WalrusVault API - Decentralized file storage with SUI wallet auth
 * OpenAPI spec version: 0.1.0
 */

export interface StorageStats {
  totalFiles: number;
  totalSize: number;
  totalSizeMb: number;
  /** @nullable */
  oldestUpload: string | null;
  /** @nullable */
  newestUpload: string | null;
  walrusFiles: number;
  localFiles: number;
}